# back/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/offers": {
      "post": {
        "operationId": "OffersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Create an offer",
        "tags": [
          "offers"
        ]
      },
      "get": {
        "description": "Most recent application first by default. Filters combine with AND; offers without a value for the sort field come last.",
        "operationId": "OffersController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of offers per page.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of offers to skip.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "integer"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text contained in the title, company or location, ignoring case. Blank is ignored.",
            "schema": {
              "maxLength": 200,
              "example": "backend",
              "type": "string"
            }
          },
          {
            "name": "appliedFrom",
            "required": false,
            "in": "query",
            "description": "Only offers applied on or after this day, `YYYY-MM-DD`.",
            "schema": {
              "format": "date",
              "example": "2026-09-01",
              "type": "string"
            }
          },
          {
            "name": "appliedTo",
            "required": false,
            "in": "query",
            "description": "Only offers applied on or before this day, `YYYY-MM-DD`. Not before `appliedFrom`.",
            "schema": {
              "format": "date",
              "example": "2026-09-30",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Only offers with one of these statuses; repeat the parameter for several (`?status=applied&status=interviewing`). All when absent.",
            "schema": {
              "example": [
                "applied",
                "interviewing"
              ],
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "applied",
                  "interviewing",
                  "offered",
                  "accepted",
                  "rejected",
                  "ghosted",
                  "withdrawn"
                ]
              }
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field. Offers with no value for it (no application date, no company) come last. Statuses sort in the order of a search, `applied` first.",
            "schema": {
              "$ref": "#/components/schemas/OfferSortField"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort direction. Default: `desc` for dates (most recent first), `asc` for text.",
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "List offers: filtered, sorted and paginated",
        "tags": [
          "offers"
        ]
      }
    },
    "/offers/{id}": {
      "get": {
        "operationId": "OffersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No offer with this id (`resource-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No offer has the id 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d. It may have been deleted; list offers (GET /offers) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "offer"
                }
              }
            }
          }
        },
        "summary": "Get an offer",
        "tags": [
          "offers"
        ]
      },
      "patch": {
        "description": "Partial update: absent fields are left unchanged, `null` clears an optional field.",
        "operationId": "OffersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfferDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No offer with this id (`resource-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No offer has the id 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d. It may have been deleted; list offers (GET /offers) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "offer"
                }
              }
            }
          }
        },
        "summary": "Update an offer",
        "tags": [
          "offers"
        ]
      },
      "delete": {
        "operationId": "OffersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Offer id.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "400": {
            "description": "Invalid request (`validation-error`): `errors` lists every invalid value of the body, query or path. Also `malformed-request` when the body is not valid JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/validation-error",
                  "title": "Invalid request",
                  "status": 400,
                  "detail": "2 values are invalid: see `errors`, fix them and send the request again.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "errors": [
                    {
                      "in": "body",
                      "name": "/title",
                      "code": "isNotEmpty",
                      "detail": "title should not be empty"
                    },
                    {
                      "in": "body",
                      "name": "/status",
                      "code": "isIn",
                      "detail": "status must be one of the following values: applied, interviewing",
                      "allowed": [
                        "applied",
                        "interviewing"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No offer with this id (`resource-not-found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                },
                "example": {
                  "type": "/problems/resource-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No offer has the id 0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d. It may have been deleted; list offers (GET /offers) to get valid ids.",
                  "instance": "/offers/0199a7a4-3c2e-7b6a-9c1d-2f3e4a5b6c7d",
                  "resource": "offer"
                }
              }
            }
          }
        },
        "summary": "Delete an offer and its interview steps",
        "tags": [
          "offers"
        ]
      }
    },
...
  },
  "info": {
    "title": "emploi API",
    "description": "Job search history: the job offers the user responded to and the steps of their interview processes.\n\nErrors are RFC 9457 Problem Details (`application/problem+json`): `type` names the kind of problem (`GET` it, e.g. `/problems/validation-error`, for what it means and how to fix the request), `detail` says what happened and what to do, and a `validation-error` lists every invalid value in `errors`, with its location, the rule broken and, when they help, the accepted values.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "InvalidParamDto": {
        "type": "object",
        "properties": {
          "in": {
            "type": "string",
            "enum": [
              "body",
              "query",
              "path"
            ]
          },
          "name": {
            "type": "string",
            "description": "JSON Pointer (RFC 6901) into the body, e.g. `/title`, or the query or path parameter name.",
            "example": "/title"
          },
          "code": {
            "type": "string",
            "description": "Rule broken, stable: e.g. `isNotEmpty`, `maxLength`, `isIn`, `isUuid`, `whitelistValidation` (unknown field: remove it).",
            "example": "isNotEmpty"
          },
          "detail": {
            "type": "string",
            "example": "title should not be empty"
          },
          "allowed": {
            "description": "For `isIn`: the accepted values.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxLength": {
            "type": "integer",
            "description": "For `maxLength`: the maximum number of characters."
          }
        },
        "required": [
          "in",
          "name",
          "code",
          "detail"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Kind of problem: `/problems/<kind>`, relative to the API (`GET` it for what it means and how to fix the request), or `about:blank`.",
            "example": "/problems/validation-error"
          },
          "title": {
            "type": "string",
            "description": "Summary of the kind of problem."
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "What happened this time, and what to do about it."
          },
          "instance": {
            "type": "string",
            "description": "Path of the request that failed."
          },
          "errors": {
            "description": "`validation-error`: every invalid value.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParamDto"
            }
          },
          "resource": {
            "type": "string",
            "enum": [
              "offer",
              "interview-step"
            ],
            "description": "`resource-not-found`: the kind of resource not found."
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "OfferStatus": {
        "type": "string",
        "enum": [
          "applied",
          "interviewing",
          "offered",
          "accepted",
          "rejected",
          "ghosted",
          "withdrawn"
        ],
        "description": "Where the application stands, chosen by the user: `applied`: sent, no answer yet; `interviewing`: interviews under way; `offered`: the company made an offer; `accepted`: the user accepted it; `rejected`: the company said no; `ghosted`: no answer after a long time; `withdrawn`: the user withdrew."
      },
      "CreateOfferDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Job title. Trimmed.",
            "maxLength": 200,
            "example": "Développeur backend"
          },
          "company": {
            "type": "string",
            "nullable": true,
            "description": "Employer, when the offer names it. Blank is stored as `null`.",
            "maxLength": 200,
            "example": "Acme"
          },
          "url": {
            "type": "string",
            "nullable": true,
            "format": "uri",
            "description": "Link to the offer, `http` or `https`.",
            "maxLength": 2048,
            "example": "https://jobs.example.com/42"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "City, country or \"Remote\".",
            "maxLength": 200,
            "example": "Lyon"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Text of the offer.",
            "maxLength": 20000
          },
          "appliedAt": {
            "type": "string",
            "nullable": true,
            "format": "date",
            "description": "Day the user responded to the offer, `YYYY-MM-DD`.",
            "example": "2026-09-28"
          },
          "status": {
            "description": "Where the application stands, chosen by the user: `applied`: sent, no answer yet; `interviewing`: interviews under way; `offered`: the company made an offer; `accepted`: the user accepted it; `rejected`: the company said no; `ghosted`: no answer after a long time; `withdrawn`: the user withdrew.",
            "example": "interviewing",
            "default": "applied",
            "allOf": [
              {
                "$ref": "#/components/schemas/OfferStatus"
              }
            ]
          }
        },
        "required": [
          "title"
        ]
      },
      "OfferDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Job title. Trimmed.",
            "maxLength": 200,
            "example": "Développeur backend"
          },
          "company": {
            "type": "string",
            "nullable": true,
            "description": "Employer, when the offer names it. Blank is stored as `null`.",
            "maxLength": 200,
            "example": "Acme"
          },
          "url": {
            "type": "string",
            "nullable": true,
            "format": "uri",
            "description": "Link to the offer, `http` or `https`.",
            "maxLength": 2048,
            "example": "https://jobs.example.com/42"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "City, country or \"Remote\".",
            "maxLength": 200,
            "example": "Lyon"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Text of the offer.",
            "maxLength": 20000
          },
          "appliedAt": {
            "type": "string",
            "nullable": true,
            "format": "date",
            "description": "Day the user responded to the offer, `YYYY-MM-DD`.",
            "example": "2026-09-28"
          },
          "status": {
            "description": "Where the application stands, chosen by the user: `applied`: sent, no answer yet; `interviewing`: interviews under way; `offered`: the company made an offer; `accepted`: the user accepted it; `rejected`: the company said no; `ghosted`: no answer after a long time; `withdrawn`: the user withdrew.",
            "example": "interviewing",
            "allOf": [
              {
                "$ref": "#/components/schemas/OfferStatus"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-01T08:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T09:45:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "url",
          "location",
          "description",
          "appliedAt",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "OfferSortField": {
        "type": "string",
        "enum": [
          "appliedAt",
          "createdAt",
          "title",
          "company",
          "status"
        ]
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
      "OfferPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Newest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfferDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Number of offers across all pages."
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "UpdateOfferDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Job title. Trimmed; can be changed but not cleared.",
            "maxLength": 200,
            "example": "Développeur backend"
          },
          "company": {
            "type": "string",
            "nullable": true,
            "description": "Employer, when the offer names it. Blank is stored as `null`.",
            "maxLength": 200,
            "example": "Acme"
          },
          "url": {
            "type": "string",
            "nullable": true,
            "format": "uri",
            "description": "Link to the offer, `http` or `https`.",
            "maxLength": 2048,
            "example": "https://jobs.example.com/42"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "City, country or \"Remote\".",
            "maxLength": 200,
            "example": "Lyon"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Text of the offer.",
            "maxLength": 20000
          },
          "appliedAt": {
            "type": "string",
            "nullable": true,
            "format": "date",
            "description": "Day the user responded to the offer, `YYYY-MM-DD`.",
            "example": "2026-09-28"
          },
          "status": {
            "description": "Where the application stands, chosen by the user: `applied`: sent, no answer yet; `interviewing`: interviews under way; `offered`: the company made an offer; `accepted`: the user accepted it; `rejected`: the company said no; `ghosted`: no answer after a long time; `withdrawn`: the user withdrew.",
            "example": "interviewing",
            "allOf": [
              {
                "$ref": "#/components/schemas/OfferStatus"
              }
            ]
          }
        }
      },
...
    }
  }
}